backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.0
pandas==2.2.0
gspread==6.0.0
oauth2client==4.1.3
openpyxl==3.1.2
google-genai
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==3.2.2